schema update, flyway version up

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,7 +1,7 @@
 drop table if exists account;
 create table account
 (
-    id       int(10) not null auto_increment,
+    id       bigint not null auto_increment,
     password varchar(255),
     username varchar(255),
     primary key (id)
